Storyboard
These are different screenshots show how the game is going to look like.
In the meantime the main menu, control menu part 1 game level (medium
mode) are being explained while in order to show what could be the idea
behind the game. The others are different screenshots of what happens
once the game has been lost or not.